backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2
pdf2image==1.17.0
Pillow==10.4.0
aiofiles==24.1.0
python-dateutil==2.9.0
